+++ backend/libs/shared/common/types/src/index.ts
@@ -3,4 +3,5 @@ export * from './lib/identifiers.dto';
 export * from './lib/jokes.dto';
 export * from './lib/keystore.dto';
 export * from './lib/ocr-space.dto';
+export * from './lib/spotify.dto';
 export * from './lib/todo.dto';